+++ src/providers/data/MemoryDataProvider.ts
@@ -37,8 +37,8 @@ export class MemoryDataProvider implements IDataProvider.IDataProvider {
 
     Options = new CommonSqlDataProviderOptions()
 
-    constructor(sourceName: string, sourceParams: TConfigSource) {
-        this.SourceName = sourceName
+    constructor(source: string, sourceParams: TConfigSource) {
+        this.SourceName = source
         this.Init(sourceParams)
         this.Connect()
     }
@@ -129,7 +129,7 @@ export class MemoryDataProvider implements IDataProvider.IDataProvider {
             if (options?.Cache)
                 Cache.Set({
                     ...schemaRequest,
-                    sourceName: this.SourceName
+                    source: this.SourceName
                 },
                     data
                 )